Fixed Scrolling issues on large dataset
On large dataset the snapToItem Method did not correctly snap to the index. Since FlatList now includes a snapToIndex method it is now used instead of snapToOffset. In addtion the values for maxToRenderPerBatch, initialNumToRender and getItemLayout are only automatically overwritten if they are not given in the props to make the list easier customizable. This also helps on the scrolling issue.